vcmi: split IBonusBearer
