+++ lib/BasicTypes.cpp
@@ -13,6 +13,7 @@
 #include "VCMI_Lib.h"
 #include "GameConstants.h"
 #include "bonuses/HeroBonus.h"
+#include "bonuses/IBonusBearer.h"
 
 #include <vcmi/Creature.h>
 #include <vcmi/Faction.h>
